Refactor for easier extension of functionality
